refactor: rename trello functions script
